# src/main/resources/openapi.yaml
openapi: 3.0.2
info:
  description: Build Once. Integrate Your Blockchain Everywhere.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Rosetta
  version: 1.4.10
servers:
- url: /
paths:
  /network/list:
    post:
      description: This endpoint returns a list of NetworkIdentifiers that the Rosetta
        server supports.
      operationId: networkList
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MetadataRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkListResponse'
          description: Expected response to a valid request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Get List of Available Networks
      tags:
      - Network
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Network
  /network/status:
    post:
      description: This endpoint returns the current status of the network requested.
        Any NetworkIdentifier returned by /network/list should be accessible here.
      operationId: networkStatus
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkStatusResponse'
          description: Expected response to a valid request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Get Network Status
      tags:
      - Network
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Network
  /network/options:
    post:
      description: "This endpoint returns the version information and allowed network-specific\
        \ types for a NetworkIdentifier. Any NetworkIdentifier returned by /network/list\
        \ should be accessible here. Because options are retrievable in the context\
        \ of a NetworkIdentifier, it is possible to define unique options for each\
        \ network."
      operationId: networkOptions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkOptionsResponse'
          description: Expected response to a valid request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Get Network Options
      tags:
      - Network
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Network
...
  /events/blocks:
    post:
      description: "`/events/blocks` allows the caller to query a sequence of BlockEvents\
        \ indicating which blocks were added and removed from storage to reach the\
        \ current state. Following BlockEvents allows lightweight clients to update\
        \ their state without needing to implement their own syncing logic (like finding\
        \ the common parent in a reorg). `/events/blocks` is considered an \"indexer\"\
        \ endpoint and Rosetta implementations are not required to complete it to\
        \ adhere to the Rosetta spec. However, any Rosetta \"indexer\" MUST support\
        \ this endpoint."
      operationId: eventsBlocks
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventsBlocksRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsBlocksResponse'
          description: Expected response to a valid request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: "[INDEXER] Get a range of BlockEvents"
      tags:
      - Events
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Events
...
components:
  schemas:
    NetworkIdentifier:
      description: The network_identifier specifies which network a particular object
        is associated with.
      example:
        blockchain: bitcoin
        sub_network_identifier:
          metadata:
            producer: 0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5
          network: shard 1
        network: mainnet
      properties:
        blockchain:
          example: bitcoin
          type: string
        network:
          description: "If a blockchain has a specific chain-id or network identifier,\
            \ it should go in this field. It is up to the client to determine which\
            \ network-specific identifier is mainnet, preprod, or preview."
          example: mainnet
          type: string
        sub_network_identifier:
          $ref: '#/components/schemas/SubNetworkIdentifier'
      required:
      - blockchain
      - network
      type: object
...
    BlockIdentifier:
      description: The block_identifier uniquely identifies a block in a particular
        network.
      example:
        index: 1123941
        hash: 0x1f2cc6c5027d2f201a5453ad1119574d2aed23a392654742ac3c78783c071f85
      properties:
        index:
          description: This is also known as the block height.
          example: 1123941
          format: int64
          type: integer
        hash:
          example: 0x1f2cc6c5027d2f201a5453ad1119574d2aed23a392654742ac3c78783c071f85
          type: string
      required:
      - hash
      - index
      type: object
...
    Peer:
      description: A Peer is a representation of a node's peer.
      example:
        metadata: "{}"
        peer_id: 0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5
      properties:
        peer_id:
          example: 0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5
          type: string
        metadata:
          type: object
      required:
      - peer_id
      type: object
...
    BlockEvent:
      description: BlockEvent represents the addition or removal of a BlockIdentifier
        from storage. Streaming BlockEvents allows lightweight clients to update their
        own state without needing to implement their own syncing logic.
      example:
        sequence: 5
        type: null
        block_identifier:
          index: 1123941
          hash: 0x1f2cc6c5027d2f201a5453ad1119574d2aed23a392654742ac3c78783c071f85
      properties:
        sequence:
          description: sequence is the unique identifier of a BlockEvent within the
            context of a NetworkIdentifier.
          example: 5
          format: int64
          minimum: 0
          type: integer
        block_identifier:
          $ref: '#/components/schemas/BlockIdentifier'
        type:
          $ref: '#/components/schemas/BlockEventType'
      required:
      - block_identifier
      - sequence
      - type
      type: object
    BlockEventType:
      description: BlockEventType determines if a BlockEvent represents the addition
        or removal of a block.
      enum:
      - block_added
      - block_removed
      type: string
...
    MetadataRequest:
      description: A MetadataRequest is utilized in any request where the only argument
        is optional metadata.
      example:
        metadata: "{}"
      properties:
        metadata:
          type: object
      type: object
    NetworkListResponse:
      description: A NetworkListResponse contains all NetworkIdentifiers that the
        node can serve information for.
      example:
        network_identifiers:
        - blockchain: bitcoin
          sub_network_identifier:
            metadata:
              producer: 0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5
            network: shard 1
          network: mainnet
        - blockchain: bitcoin
          sub_network_identifier:
            metadata:
              producer: 0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5
            network: shard 1
          network: mainnet
      properties:
        network_identifiers:
          items:
            $ref: '#/components/schemas/NetworkIdentifier'
          type: array
      required:
      - network_identifiers
      type: object
    NetworkRequest:
      description: A NetworkRequest is utilized to retrieve some data specific exclusively
        to a NetworkIdentifier.
      example:
        metadata: "{}"
        network_identifier:
          blockchain: bitcoin
          sub_network_identifier:
            metadata:
              producer: 0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5
            network: shard 1
          network: mainnet
      properties:
        network_identifier:
          $ref: '#/components/schemas/NetworkIdentifier'
        metadata:
          type: object
      required:
      - network_identifier
      type: object
    NetworkStatusResponse:
      description: "NetworkStatusResponse contains basic information about the node's\
        \ view of a blockchain network. It is assumed that any BlockIdentifier.Index\
        \ less than or equal to CurrentBlockIdentifier.Index can be queried. If a\
        \ Rosetta implementation prunes historical state, it should populate the optional\
        \ `oldest_block_identifier` field with the oldest block available to query.\
        \ If this is not populated, it is assumed that the `genesis_block_identifier`\
        \ is the oldest queryable block. If a Rosetta implementation performs some\
        \ pre-sync before it is possible to query blocks, sync_status should be populated\
        \ so that clients can still monitor healthiness. Without this field, it may\
        \ appear that the implementation is stuck syncing and needs to be terminated."
      example:
        oldest_block_identifier:
          index: 1123941
          hash: 0x1f2cc6c5027d2f201a5453ad1119574d2aed23a392654742ac3c78783c071f85
        current_block_identifier:
          index: 1123941
          hash: 0x1f2cc6c5027d2f201a5453ad1119574d2aed23a392654742ac3c78783c071f85
        peers:
        - metadata: "{}"
          peer_id: 0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5
        - metadata: "{}"
          peer_id: 0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5
        sync_status:
          target_index: 150
          synced: true
          stage: header sync
          current_index: 100
        current_block_timestamp: 1582833600000
        genesis_block_identifier:
          index: 1123941
          hash: 0x1f2cc6c5027d2f201a5453ad1119574d2aed23a392654742ac3c78783c071f85
      properties:
        current_block_identifier:
          $ref: '#/components/schemas/BlockIdentifier'
        current_block_timestamp:
          description: The timestamp of the block in milliseconds since the Unix Epoch.
            The timestamp is stored in milliseconds because some blockchains produce
            blocks more often than once a second.
          example: 1582833600000
          format: int64
          minimum: 0
          type: integer
        genesis_block_identifier:
          $ref: '#/components/schemas/BlockIdentifier'
        oldest_block_identifier:
          $ref: '#/components/schemas/BlockIdentifier'
        sync_status:
          $ref: '#/components/schemas/SyncStatus'
        peers:
          items:
            $ref: '#/components/schemas/Peer'
          type: array
      required:
      - current_block_identifier
      - current_block_timestamp
      - genesis_block_identifier
      - peers
      type: object
    NetworkOptionsResponse:
      description: "NetworkOptionsResponse contains information about the versioning\
        \ of the node and the allowed operation statuses, operation types, and errors."
      example:
        allow:
          balance_exemptions:
          - sub_account_address: staking
            currency:
              symbol: BTC
              metadata: ""
              decimals: 8
            exemption_type: null
          - sub_account_address: staking
            currency:
              symbol: BTC
              metadata: ""
              decimals: 8
            exemption_type: null
          mempool_coins: true
          operation_types:
          - TRANSFER
          - TRANSFER
          historical_balance_lookup: true
          operation_statuses:
          - status: SUCCESS
            successful: true
          - status: SUCCESS
            successful: true
          timestamp_start_index: 0
          call_methods:
          - eth_call
          - eth_call
          errors:
          - retriable: true
            code: 12
            description: This error is returned when the requested AccountIdentifier
              is improperly formatted.
            details:
              address: 0x1dcc4de8dec75d7aab85b567b6
              error: not base64
            message: Invalid account format
          - retriable: true
            code: 12
            description: This error is returned when the requested AccountIdentifier
              is improperly formatted.
            details:
              address: 0x1dcc4de8dec75d7aab85b567b6
              error: not base64
            message: Invalid account format
        version:
          metadata: "{}"
          rosetta_version: 1.2.5
          node_version: 1.0.2
          middleware_version: 0.2.7
      properties:
        version:
          $ref: '#/components/schemas/Version'
        allow:
          $ref: '#/components/schemas/Allow'
      required:
      - allow
      - version
      type: object
...
    EventsBlocksRequest:
      description: EventsBlocksRequest is utilized to fetch a sequence of BlockEvents
        indicating which blocks were added and removed from storage to reach the current
        state.
      example:
        offset: 5
        limit: 5
        network_identifier:
          blockchain: bitcoin
          sub_network_identifier:
            metadata:
              producer: 0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5
            network: shard 1
          network: mainnet
      properties:
        network_identifier:
          $ref: '#/components/schemas/NetworkIdentifier'
        offset:
          description: "offset is the offset into the event stream to sync events\
            \ from. If this field is not populated, we return the limit events backwards\
            \ from tip. If this is set to 0, we start from the beginning."
          example: 5
          format: int64
          minimum: 0
          type: integer
        limit:
          description: limit is the maximum number of events to fetch in one call.
            The implementation may return <= limit events.
          example: 5
          format: int64
          minimum: 0
          type: integer
      required:
      - network_identifier
      type: object
    EventsBlocksResponse:
      description: EventsBlocksResponse contains an ordered collection of BlockEvents
        and the max retrievable sequence.
      example:
        max_sequence: 5
        events:
        - sequence: 5
          type: null
          block_identifier:
            index: 1123941
            hash: 0x1f2cc6c5027d2f201a5453ad1119574d2aed23a392654742ac3c78783c071f85
        - sequence: 5
          type: null
          block_identifier:
            index: 1123941
            hash: 0x1f2cc6c5027d2f201a5453ad1119574d2aed23a392654742ac3c78783c071f85
      properties:
        max_sequence:
          description: max_sequence is the maximum available sequence number to fetch.
          example: 5
          format: int64
          minimum: 0
          type: integer
        events:
          description: events is an array of BlockEvents indicating the order to add
            and remove blocks to maintain a canonical view of blockchain state. Lightweight
            clients can use this event stream to update state without implementing
            their own block syncing logic.
          items:
            $ref: '#/components/schemas/BlockEvent'
          type: array
      required:
      - events
      - max_sequence
      type: object
...
    Error:
      description: "Instead of utilizing HTTP status codes to describe node errors\
        \ (which often do not have a good analog), rich errors are returned using\
        \ this object. Both the code and message fields can be individually used to\
        \ correctly identify an error. Implementations MUST use unique values for\
        \ both fields."
      example:
        retriable: true
        code: 12
        description: This error is returned when the requested AccountIdentifier is
          improperly formatted.
        details:
          address: 0x1dcc4de8dec75d7aab85b567b6
          error: not base64
        message: Invalid account format
      properties:
        code:
          description: "Code is a network-specific error code. If desired, this code\
            \ can be equivalent to an HTTP status code."
          example: 12
          format: int32
          minimum: 0
          type: integer
        message:
          description: "Message is a network-specific error message. The message MUST\
            \ NOT change for a given code. In particular, this means that any contextual\
            \ information should be included in the details field."
          example: Invalid account format
          type: string
        description:
          description: "Description allows the implementer to optionally provide additional\
            \ information about an error. In many cases, the content of this field\
            \ will be a copy-and-paste from existing developer documentation. Description\
            \ can ONLY be populated with generic information about a particular type\
            \ of error. It MUST NOT be populated with information about a particular\
            \ instantiation of an error (use `details` for this). Whereas the content\
            \ of Error.Message should stay stable across releases, the content of\
            \ Error.Description will likely change across releases (as implementers\
            \ improve error documentation). For this reason, the content in this field\
            \ is not part of any type assertion (unlike Error.Message)."
          example: This error is returned when the requested AccountIdentifier is
            improperly formatted.
          type: string
        retriable:
          description: An error is retriable if the same request may succeed if submitted
            again.
          type: boolean
        details:
          $ref: '#/components/schemas/Error_details'
      required:
      - code
      - message
      - retriable
      type: object
...
    Error_details:
      description: Often times it is useful to return context specific to the request
        that caused the error (i.e. a sample of the stack trace or impacted account)
        in addition to the standard error message.
      example:
        address: 0x1dcc4de8dec75d7aab85b567b6
        error: not base64
      properties:
        message:
          type: string
      required:
      - message
      type: object